+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Fixed
+- Flash viewer - subtract blend mode
 
 ## [16.0.4] - 2022-11-03
 ### Fixed